% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajorant.figures-package.R
\docType{package}
\name{ajorant.figures-package}
\alias{ajorant.figures}
\alias{ajorant.figures-package}
\title{ajorant.figures: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Amaury Jorant \email{amaury.jorant@protonmail.com}

}
\keyword{internal}
