% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{fig_grid}
\alias{fig_grid}
\title{Multitude de graphiques sur une page}
\usage{
fig_grid(
  list_graph,
  titre,
  fichier,
  ncol,
  .base_dim = c(3, 2),
  .rel_widths = 1,
  .rel_heights = 1,
  .noms_lignes = NULL,
  .noms_col = NULL,
  .legende = NULL,
  .L_dim = c(2, 2),
  .label_position = 1,
  .espace_insecable = "_"
)
}
\arguments{
\item{list_graph}{Liste des graphiques à intégrer à la figure.}

\item{titre}{Titre de la figure.}

\item{fichier}{Fichier de destination de la figure.}

\item{ncol}{Nombre de colones de graphiques. Le nombre de ligne est déduit du nombre de graphiques à afficher.}

\item{.base_dim}{Dimension x, y des graphiques de base. (en inches)}

\item{.rel_widths}{Vecteur, largeur relative de chaque colonne.
Le nombre d'éléments doit être égale au nombre de colonnes.}

\item{.rel_heights}{Vecteur, hauteur relative de chaque colonne.
Le nombre d'éléments doit être égale au nombre de lignes.}

\item{.noms_lignes}{Vecteur. Label des lignes de graphique.}

\item{.noms_col}{Vecteur. Label des colonnes de graphiques.}

\item{.legende}{Graphique de égende, affiché sur la même ligne que le titre.
A utiliser lorsque la légende est la même pour tous les graphiques.}

\item{.L_dim}{dimensions x, y de la légende (en inches).}

\item{.label_position}{alignement des titres de ligne et colonne: 0 à gauche, 0.5 au centre et 1 à droite.}

\item{.espace_insecable}{charactère utilisé comme espace insécable dans le titre.}
}
\value{
Graphique sauvegardé directement dans \code{fichier}.
}
\description{
Arrange et exporte plusieurs graphique en grille sur une page.
}
\section{Erreurs connues}{

une erreur a été observée dans plot_grid() pour des graph tels que :
\verb{ggplot() +}
\code{geom_point(data = df, aes(x = df[,var1], y = var2))}

=> Les points étaient mélangés d'un graph à l'autre.
Il faut privilégier \code{.data[[var]]} à \code{data.frame[[,var]]} ou \code{aes_string(var, 'RF_votes')}.
}

\section{Futurs évolutions}{

Les rel_heights et rel_widths agrandissent la taille de chaque graph.
Il faudrait modifier les dims finales du graph car pour l'instant elle sont calculées pour tous les graphs à 1.
}

