% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{graph_perf_ROC}
\alias{graph_perf_ROC}
\title{Graphique de performances ROC}
\usage{
graph_perf_ROC(data, obs, sim, seuil_obs, seuil_sim)
}
\arguments{
\item{data}{\code{data.frame} des données en entrée ...}

\item{obs}{Colonne des observés.}

\item{sim}{Colonne des simulés.}

\item{seuil_obs}{Seuil a appliquer aux observées pour séparer positif et négatif}

\item{seuil_sim}{Seuil a appliquer aux simulées pour séparer positif et négatif

Les indicateurs calculés sont :
sensibilité, spécificité, taux de bien prédit et Youden Index (sens+spe-1)
les \% dans les angles correspondent à la part dans le cadrant}
}
\value{
graphique ggplot2 avec indicateurs
}
\description{
Figure performances adapté au analyse de type ROC. Calcul des indicateurs :
sensibilité, spécificité, taux de bien prédit et Youden Index (sens+spe-1)
}
