% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{leaf_colored_markers}
\alias{leaf_colored_markers}
\title{Liste de marqueurs colorés pour Leaflet}
\usage{
leaf_colored_markers(palette)
}
\arguments{
\item{palette}{Palette de couleur sous forme \code{list(name = couleur_hexadecimal)}}
}
\value{
liste de forme \code{leaflet::iconList()}
}
\description{
Édite une liste de marqueurs colorés utilisable dans leaflet
}
\examples{

list_col <- list('rouge'="#FF0000",
                 'vert'='#C5FF00',
                 'bleu'='#00FFFF',
                 'violet'='#9100FF')

my_markers <- leaf_colored_markers(list_col)

leaflet::leaflet() |>
leaflet::addMarkers(lat = c(0,0,1,1),
                    lng = c(0,1,0,1),
                    icon = my_markers[names(list_col)])
}
