% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{mega_Palette}
\alias{mega_Palette}
\title{Palette de couleurs extra-large}
\usage{
mega_Palette(N)
}
\arguments{
\item{N}{Le nombre couleurs nécessaires dans la palette.}
}
\value{
palette de couleurs hexadécimal (vecteur).
}
\description{
Palette supportant jusqu'à 50 couleurs à peut près discernables.

Attention: Non adapté aux daltoniens.
La distinguibilité depend fortement de l'écran et des paramètres d'affichage.
}
