% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.figures-package.R
\docType{package}
\name{pack.figures-package}
\alias{pack.figures}
\alias{pack.figures-package}
\title{pack.figures: Fonctions graphiques polyvalentes}
\description{
Graphiques préconfigurés, mise en forme de graphiques, cartographie...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Amaury974/pack.figures}
  \item Report bugs at \url{https://github.com/Amaury974/pack.figures/issues}
}

}
\author{
\strong{Maintainer}: Amaury Jorant \email{amaury.jorant@protonmail.com}

}
\keyword{internal}
