% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{quel_dep}
\alias{quel_dep}
\title{Carto : dans quel département}
\usage{
quel_dep(data, long = "longitude", lat = "latitude")
}
\arguments{
\item{data}{Tableau de données.}

\item{long}{Colonne de longitude.}

\item{lat}{Colonne de lattitude.}
}
\value{
liste avec :
\itemize{
\item Le tableau \code{data} agrémenté d'une colonne \code{departement}.
\item Un tableau répertoriant les centroïdes des départements.
}
}
\description{
Identification du département à partir de coordonnées GPS.
}
