% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{rf_fig}
\alias{rf_fig}
\title{Figure RandomForest}
\usage{
rf_fig(random_forest, titre = names(random_forest), fichier, .expl_var = FALSE)
}
\arguments{
\item{random_forest}{Résultat de l'exécution de la fonction randomforest du package RandomForest.}

\item{titre}{Tite du Graphique.}

\item{fichier}{Fichier ou enregistre la figure.}

\item{.expl_var}{Doit on afficher le bas de page explicatif des variables représenté ?}
}
\value{
Sauvegarde dan \code{fichier} la figure d'analyse de la RandomForest.
}
\description{
Génère et enregistre la comparaison des \%IncMSE et des IncNodePurity.
}
