% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajorant.figures-package.R
\docType{package}
\name{ajorant.figures-package}
\alias{ajorant.figures}
\alias{ajorant.figures-package}
\title{ajorant.figures: Fonctions graphiques polyvalentes}
\description{
Graphiques préconfigurés, mise en forme de graphiques, cartographie...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Amaury974/ajorant.figures}
  \item Report bugs at \url{https://github.com/Amaury974/ajorant.figures/issues}
}

}
\author{
\strong{Maintainer}: Amaury Jorant \email{amaury.jorant@protonmail.com}

}
\keyword{internal}
